requirement.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
decorator==5.1.1
Django==4.1.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-jalali==6.0.0
django-multiupload==0.6.1
django-ranged-response==0.2.0
django-sendfile2==0.7.0
django-simple-captcha==0.5.17
django-widget-tweaks==1.4.12
gunicorn==20.1.0
jdatetime==4.1.0
Pillow==9.3.0
psycopg2-binary==2.9.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
self==2020.12.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
